FreeBSD system calls
freebsd syscalls are identical to linux ones (it seems) except for the syscall numbers

These are written as "open_syscall" instead of "open_linux" so user mode code doesn't have to know if we
run linux or freebsd. Linux syscalls have not yet been converted, but they will be.

TODO: Makefile rules to select either freebsd or linux syscalls